Fix payload buffer overflow and four other latent bugs
State::split wrote past the end of the 50 byte name/temperature buffers
for any payload longer than BUFFER_SIZE. PubSubClient allows 256 byte
payloads, so a long sensor name corrupted the adjacent buffer. Confirmed
against the old code under AddressSanitizer. Now truncated to fit, always
null terminated.

State::split also mis-sliced payloads containing more than one delimiter,
because foundDelimiter was reassigned on every '|' and the resulting index
arithmetic went out of step. Only the first delimiter splits now.

MQTT tracked subscriptions in std::set<char *>, which deduplicated by
pointer rather than by topic text, and stored pointers it did not own.
Now std::set<std::string>.

MQTT::loop inserted into _subscriptionTopics while iterating it during
reconnect; it now re-subscribes through _client directly.

Both reconnect back-offs compared signed long against millis(), which
breaks at the ~49 day rollover and can wedge reconnection. Now uint32_t.

StickyWiFi.cpp included "StickyWifi.h" with the wrong case, which only
resolves on case-insensitive filesystems.

platformio.ini: drop the hardcoded Windows upload_port so the port is
auto-detected on macOS, and set monitor_speed.

Verified with `pio run` for m5stack-core2, and split() covered by a host
test over the README payload formats plus over-length input.

+++ src/StickyWiFi.cpp
@@ -1,8 +1,8 @@
-#include "StickyWifi.h"
+#include "StickyWiFi.h"
 #include <WiFi.h>
 #include <M5Unified.h>
 
-const long backOffTimeout = 5000;
+const uint32_t backOffTimeout = 5000;
 
 StickyWiFi::StickyWiFi()
 {
@@ -22,7 +22,9 @@ wl_status_t StickyWiFi::init(char *ssid, char *passphrase)
 
 wl_status_t StickyWiFi::loop()
 {
-  long now = millis();
+  // Unsigned, so the subtraction below stays correct across the millis()
+  // rollover at ~49 days of uptime.
+  uint32_t now = millis();
   _status = WiFi.status();
   if (_status == WL_CONNECTED)
   {
